\name{Dc}
\alias{Dc}\title{ Matriz diagonal con entradas decimales}
\description{ Función para crear la matriz diagonal con entradas enteras aleatorias
}
\usage{ Dc(n, a, b,d) }
\arguments{
\item{n}{ Tamaño de matriz }
\item{a}{ Límite inferior del intervalo}
\item{b}{ Límite superior del intervalo}
\item{d}{ Dígitos decimales}}
\examples{D=Dc(3,-1,10,2) }
