\name{MTi}
\alias{MTi}
\title{ Matriz triangular superior con entradas aleatoria enteras}
\description{ Función para crear una matriz superior aleatoria}
\usage{ MTi(n, a, b) }
\arguments{
  \item{n}{ Tamaño de matriz}
  \item{a}{ Límite inferior del intervalo}
  \item{b}{ Límite superior del intervalo}
  }
\examples{M=MTi(4,-15,5)
}
