\name{MTic}
\alias{MTic}\title{ Matriz triangular inferior con entradas aleatorias}
\description{ Función para crear una matriz superior aleatoria}
\usage{ MTic(n, a, b, d) }
\arguments{
  \item{n}{ Tamaño de la matriz}
  \item{a}{ Límite inferior de la matriz}
  \item{b}{ Límite superior de la matriz}
  \item{d}{ Dígitos decimales}
}
\examples{
M=MTic(4,-5,5,2)
}
