\name{MTs}
\alias{MTs}
\title{ Matriz triangular superior con entradas aleatoria enteras}
\description{ Función para crear una matriz superior aleatoria}
\usage{ MTs(n, a, b) }
\arguments{
  \item{n}{ Tamaño de matriz}
  \item{a}{ Límite inferior del intervalo}
  \item{b}{ Límite superior del intervalo}
}
\examples{
M=MTs(4,-5,5)
}
