\name{MTsc}
\alias{MTsc}\title{  Matriz triangular superior con entradas aleatorias}
\description{ Función para crear una matriz superior aleatoria}
\usage{ MTsc(n, a, b, d) }
\arguments{
  \item{n}{ Tamaño de la matriz}
  \item{a}{ Límite inferior del intervalo}
  \item{b}{ Límite superior del intervalo}
  \item{d}{ Dígitos decimales}
}
\examples{
M=MTsc(4,-5,5,2)
}
